pyproject.toml
[tool.poetry]
name = "goodauldbooks-fastapi-sqlalchemy"
version = "0.1.0"
description = ""
authors = ["Olivier Philippon <olivier@rougemine.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
typer = "^0.4.1"
SQLAlchemy = "^1.4.36"
aiosqlite = "^0.17.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
python-slugify = "^6.1.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.950"

[tool.black]
line-length = 119
target-version = ['py310']

[tool.isort]
py_version = 310
profile = "black"
line_length = 119
multi_line_output = 3
known_first_party = ["apps"]
combine_as_imports = true

[tool.mypy]
python_version = "3.10"
mypy_path = "src/"
[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
